Move patterns into their own module.
