+++ facet.cabal
@@ -75,6 +75,7 @@ library
     Facet.CLI
     Facet.Context
     Facet.Core.Module
+    Facet.Core.Pattern
     Facet.Core.Term
     Facet.Core.Type
     Facet.Diff
